rename _state_tensor.py to _tensornet.py
