+++ tests/lightning_tensor/test_gates_and_expval.py
@@ -27,7 +27,7 @@ if device_name != "lightning.tensor":
 else:
     from pennylane_lightning.lightning_tensor import LightningTensor
     from pennylane_lightning.lightning_tensor._measurements import LightningTensorMeasurements
-    from pennylane_lightning.lightning_tensor._state_tensor import LightningTensorNet
+    from pennylane_lightning.lightning_tensor._tensornet import LightningTensorNet
 
 if not LightningDevice._new_API:  # pylint: disable=protected-access
     pytest.skip("Exclusive tests for new API. Skipping.", allow_module_level=True)
