+++ tests/lightning_tensor/test_measurements_class.py
@@ -25,7 +25,7 @@ if device_name != "lightning.tensor":
     )
 else:
     from pennylane_lightning.lightning_tensor._measurements import LightningTensorMeasurements
-    from pennylane_lightning.lightning_tensor._state_tensor import LightningTensorNet
+    from pennylane_lightning.lightning_tensor._tensornet import LightningTensorNet
 
 if not LightningDevice._CPP_BINARY_AVAILABLE:  # pylint: disable=protected-access
     pytest.skip("No binary module found. Skipping.", allow_module_level=True)
